requirements/prod.txt
python-dateutil >= 2.1
requests >= 2.4.3
pyyaml >= 3.10
simplejson >= 3.1.0
